pyproject.toml
[project]
name = "mcp-trilium"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.13.1",
    "markdown>=3.10",
    "markdownify>=1.2.2",
    "mcp>=1.21.2",
    "python-dotenv>=1.0.0",
    "requests>=2.32.5",
]
